3619: Rename SmartAttributeEditor files
