+++ common/src/View/SmartPropertyEditor.h
